+++ Assets/UGF.Utf8Json.Runtime.Tests/UGF.Utf8Json.Runtime.Tests.Utf8Json.Generated.cs
@@ -1,6 +1,8 @@
 // THIS IS A GENERATED CODE. DO NOT EDIT.
 // ReSharper disable all
 
+using Utf8Json;
+
 namespace UGF.Utf8Json.Runtime.Tests.Resolvers
 {
     using System;
@@ -18,6 +20,11 @@ namespace UGF.Utf8Json.Runtime.Tests.Resolvers
         public global::Utf8Json.IJsonFormatter<T> GetFormatter<T>()
         {
             return FormatterCache<T>.formatter;
+        }
+
+        public IJsonFormatter GetFormatter(Type type)
+        {
+            return (IJsonFormatter)UGFUtf8JsonRuntimeTestsResolverGetFormatterHelper.GetFormatter(type);
         }
 
         static class FormatterCache<T>
@@ -96,7 +103,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UnityEngine
         {
             writer.WriteInt32((Int32)value);
         }
-        
+
         public global::UnityEngine.HideFlags Deserialize(ref global::Utf8Json.JsonReader reader, global::Utf8Json.IJsonFormatterResolver formatterResolver)
         {
             return (global::UnityEngine.HideFlags)reader.ReadInt32();
@@ -138,7 +145,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
             this.____stringByteKeys = new byte[][]
             {
                 global::Utf8Json.JsonWriter.GetEncodedPropertyNameWithBeginObject("boolValue"),
-                
+
             };
         }
 
@@ -149,11 +156,11 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
                 writer.WriteNull();
                 return;
             }
-            
+
 
             writer.WriteRaw(this.____stringByteKeys[0]);
             writer.WriteBoolean(value.BoolValue);
-            
+
             writer.WriteEndObject();
         }
 
@@ -163,7 +170,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
             {
                 return null;
             }
-            
+
 
             var __BoolValue__ = default(bool);
             var __BoolValue__b__ = false;
@@ -218,7 +225,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
             this.____stringByteKeys = new byte[][]
             {
                 global::Utf8Json.JsonWriter.GetEncodedPropertyNameWithBeginObject("intValue"),
-                
+
             };
         }
 
@@ -229,11 +236,11 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
                 writer.WriteNull();
                 return;
             }
-            
+
 
             writer.WriteRaw(this.____stringByteKeys[0]);
             writer.WriteInt32(value.IntValue);
-            
+
             writer.WriteEndObject();
         }
 
@@ -243,7 +250,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
             {
                 return null;
             }
-            
+
 
             var __IntValue__ = default(int);
             var __IntValue__b__ = false;
@@ -298,7 +305,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
             this.____stringByteKeys = new byte[][]
             {
                 global::Utf8Json.JsonWriter.GetEncodedPropertyNameWithBeginObject("boolValue"),
-                
+
             };
         }
 
@@ -309,11 +316,11 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
                 writer.WriteNull();
                 return;
             }
-            
+
 
             writer.WriteRaw(this.____stringByteKeys[0]);
             writer.WriteBoolean(value.BoolValue);
-            
+
             writer.WriteEndObject();
         }
 
@@ -323,7 +330,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
             {
                 return null;
             }
-            
+
 
             var __BoolValue__ = default(bool);
             var __BoolValue__b__ = false;
@@ -378,7 +385,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
             this.____stringByteKeys = new byte[][]
             {
                 global::Utf8Json.JsonWriter.GetEncodedPropertyNameWithBeginObject("intValue"),
-                
+
             };
         }
 
@@ -389,11 +396,11 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
                 writer.WriteNull();
                 return;
             }
-            
+
 
             writer.WriteRaw(this.____stringByteKeys[0]);
             writer.WriteInt32(value.IntValue);
-            
+
             writer.WriteEndObject();
         }
 
@@ -403,7 +410,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
             {
                 return null;
             }
-            
+
 
             var __IntValue__ = default(int);
             var __IntValue__b__ = false;
@@ -458,7 +465,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
             this.____stringByteKeys = new byte[][]
             {
                 global::Utf8Json.JsonWriter.GetEncodedPropertyNameWithBeginObject("value"),
-                
+
             };
         }
 
@@ -469,11 +476,11 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
                 writer.WriteNull();
                 return;
             }
-            
+
 
             writer.WriteRaw(this.____stringByteKeys[0]);
             writer.WriteInt32(value.Value);
-            
+
             writer.WriteEndObject();
         }
 
@@ -483,7 +490,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests.Forma
             {
                 return null;
             }
-            
+
 
             var __Value__ = default(int);
             var __Value__b__ = false;
@@ -564,7 +571,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests
                 global::Utf8Json.JsonWriter.GetEncodedPropertyNameWithPrefixValueSeparator("floatValue"),
                 global::Utf8Json.JsonWriter.GetEncodedPropertyNameWithPrefixValueSeparator("intValue"),
                 global::Utf8Json.JsonWriter.GetEncodedPropertyNameWithPrefixValueSeparator("flags"),
-                
+
             };
         }
 
@@ -575,7 +582,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests
                 writer.WriteNull();
                 return;
             }
-            
+
 
             writer.WriteRaw(this.____stringByteKeys[0]);
             writer.WriteString(value.Name);
@@ -587,7 +594,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests
             writer.WriteInt32(value.IntValue);
             writer.WriteRaw(this.____stringByteKeys[4]);
             global::Utf8Json.JsonFormatterResolverExtensions.GetFormatterWithVerify<global::UnityEngine.HideFlags>(formatterResolver).Serialize(ref writer, value.Flags, formatterResolver);
-            
+
             writer.WriteEndObject();
         }
 
@@ -597,7 +604,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests
             {
                 return null;
             }
-            
+
 
             var __Name__ = default(string);
             var __Name__b__ = false;
@@ -682,7 +689,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests
             {
                 global::Utf8Json.JsonWriter.GetEncodedPropertyNameWithBeginObject("vector2"),
                 global::Utf8Json.JsonWriter.GetEncodedPropertyNameWithPrefixValueSeparator("bounds"),
-                
+
             };
         }
 
@@ -693,13 +700,13 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests
                 writer.WriteNull();
                 return;
             }
-            
+
 
             writer.WriteRaw(this.____stringByteKeys[0]);
             global::Utf8Json.JsonFormatterResolverExtensions.GetFormatterWithVerify<global::UnityEngine.Vector2>(formatterResolver).Serialize(ref writer, value.Vector2, formatterResolver);
             writer.WriteRaw(this.____stringByteKeys[1]);
             global::Utf8Json.JsonFormatterResolverExtensions.GetFormatterWithVerify<global::UnityEngine.Bounds>(formatterResolver).Serialize(ref writer, value.Bounds, formatterResolver);
-            
+
             writer.WriteEndObject();
         }
 
@@ -709,7 +716,7 @@ namespace UGF.Utf8Json.Runtime.Tests.Formatters.UGF.Utf8Json.Runtime.Tests
             {
                 return null;
             }
-            
+
 
             var __Vector2__ = default(global::UnityEngine.Vector2);
             var __Vector2__b__ = false;