+++ Packages/UGF.Utf8Json/Runtime/Plugins/Utf8Json/IJsonFormatter.cs
@@ -24,6 +24,26 @@ namespace Utf8Json
         T DeserializeFromPropertyName(ref JsonReader reader, IJsonFormatterResolver formatterResolver);
     }
 
+    public abstract class JsonFormatterBase<T> : JsonFormatterAbstract<T>, IJsonFormatter<T>
+    {
+    }
+
+    public abstract class JsonFormatterAbstract<T> : IJsonFormatter<object>
+    {
+        public abstract void Serialize(ref JsonWriter writer, T value, IJsonFormatterResolver formatterResolver);
+        public abstract T Deserialize(ref JsonReader reader, IJsonFormatterResolver formatterResolver);
+
+        void IJsonFormatter<object>.Serialize(ref JsonWriter writer, object value, IJsonFormatterResolver formatterResolver)
+        {
+            Serialize(ref writer, (T)value, formatterResolver);
+        }
+
+        object IJsonFormatter<object>.Deserialize(ref JsonReader reader, IJsonFormatterResolver formatterResolver)
+        {
+            return Deserialize(ref reader, formatterResolver);
+        }
+    }
+
     public static class JsonFormatterExtensions
     {
         public static string ToJsonString<T>(this IJsonFormatter<T> formatter, T value, IJsonFormatterResolver formatterResolver)